Decompose ModelAlicorn into its constituents
