+++ src/main/java/com/minelittlepony/model/capabilities/IModelPegasus.java
@@ -4,10 +4,7 @@ public interface IModelPegasus extends IModel {
     /**
      * Returns true if the wings are spread.
      */
-    boolean wingsAreOpen();
-
-    /**
-     * Returns true if this model is being applied to a race that has wings.
-     */
-    boolean canFly();
+    default boolean wingsAreOpen() {
+        return isFlying() || isCrouching();
+    }
 }
